Rename migration to dataflows
Jira ticket: CAMS-421
