+++ .github/workflows/continuous-deployment.yml
@@ -108,7 +108,7 @@ jobs:
       camsBasePath: ${{ vars.CAMS_BASE_PATH }}
       webappName: ${{ needs.setup.outputs.webappName }}
       apiFunctionName: ${{ needs.setup.outputs.apiFunctionName }}
-      migrationFunctionName: ${{ needs.setup.outputs.migrationFunctionName }}
+      dataflowsFunctionName: ${{ needs.setup.outputs.dataflowsFunctionName }}
       environment: ${{ needs.setup.outputs.ghaEnvironment }}
       launchDarklyEnvironment: ${{ vars.CAMS_LAUNCH_DARKLY_ENV }}
       azResourceGrpAppEncrypted: ${{ needs.setup.outputs.azResourceGrpAppEncrypted }}
@@ -133,7 +133,7 @@ jobs:
       azResourceGrpNetworkEncrypted: ${{ needs.setup.outputs.azResourceGrpNetworkEncrypted }}
       stackName: ${{ needs.setup.outputs.stackName }}
       apiFunctionName: ${{ needs.setup.outputs.apiFunctionName }}
-      migrationFunctionName: ${{ needs.setup.outputs.migrationFunctionName }}
+      dataflowsFunctionName: ${{ needs.setup.outputs.dataflowsFunctionName }}
       webappName: ${{ needs.setup.outputs.webappName }}
       environmentHash: ${{ needs.setup.outputs.environmentHash }}
       deployVnet: ${{ needs.setup.outputs.deployVnet }}
@@ -151,7 +151,7 @@ jobs:
       azResourceGrpAppEncrypted: ${{ needs.setup.outputs.azResourceGrpAppEncrypted }}
       stackName: ${{ needs.setup.outputs.stackName }}
       apiFunctionName: ${{ needs.setup.outputs.apiFunctionName }}
-      migrationFunctionName: ${{ needs.setup.outputs.migrationFunctionName }}
+      dataflowsFunctionName: ${{ needs.setup.outputs.dataflowsFunctionName }}
       webappName: ${{ needs.setup.outputs.webappName }}
       environmentHash: ${{ needs.setup.outputs.environmentHash }}
       slotName: ${{ needs.setup.outputs.slotName }}
