+++ .github/workflows/sub-build.yml
@@ -9,7 +9,7 @@ on:
       apiFunctionName:
         required: true
         type: string
-      migrationFunctionName:
+      dataflowsFunctionName:
         required: true
         type: string
       camsServerHostname:
@@ -68,13 +68,13 @@ jobs:
           npm run build-common
           npm ci
           npm run build:api
-          npm run build:migration
+          npm run build:dataflows
 
       - name: Package API
         run: OUT=${{ inputs.apiFunctionName }} npm run pack
 
       - name: Package Migration
-        run: OUT=${{ inputs.migrationFunctionName }} npm run pack:migration
+        run: OUT=${{ inputs.dataflowsFunctionName }} npm run pack:dataflows
 
       - name: Upload API Azure Functions Artifact
         uses: actions/upload-artifact@v4.4.0
@@ -86,6 +86,6 @@ jobs:
       - name: Upload Migration Azure Functions Artifact
         uses: actions/upload-artifact@v4.4.0
         with:
-          name: ${{ inputs.migrationFunctionName }}-build
-          path: backend/function-apps/migration/${{ inputs.migrationFunctionName }}.zip
+          name: ${{ inputs.dataflowsFunctionName }}-build
+          path: backend/function-apps/dataflows/${{ inputs.dataflowsFunctionName }}.zip
           if-no-files-found: error
