+++ backend/function-apps/dataflows/migration/activity/flattenBoundingArrays.ts
