+++ backend/function-apps/dataflows/package.json
@@ -1,13 +1,13 @@
 {
-  "name": "cams-pipelines",
+  "name": "cams-dataflows",
   "version": "1.0.0",
   "description": "",
-  "main": "dist/backend/function-apps/pipelines/pipelines.js",
+  "main": "dist/backend/function-apps/dataflows/dataflows.js",
   "scripts": {
     "clean": "rm -rf ./dist ./node_modules *.zip",
     "prestart": "npm run build && cp ../../.env .",
     "start": "func start --javascript",
     "build": "tsc --build tsconfig.build.json",
-    "pack": "../../pack.sh migration"
+    "pack": "../../pack.sh dataflows"
   }
 }
