+++ backend/package.json
@@ -43,22 +43,22 @@
   },
   "scripts": {
     "all": "npm run clean:all && npm ci && npm run build:all && npm run pack:all",
-    "clean:all": "npm run clean && npm run clean:api && npm run clean:pipelines",
+    "clean:all": "npm run clean && npm run clean:api && npm run clean:dataflows",
     "clean": "rm -rf ./dist && rm -rf ./node_modules && rm -rf ./coverage",
     "clean:api": "cd function-apps/api && npm run clean && cd ../..",
-    "clean:pipelines": "cd function-apps/pipelines && npm run clean && cd ../..",
-    "build:all": "npm run build-common && npm run build && npm run build:api && npm run build:pipelines",
+    "clean:dataflows": "cd function-apps/dataflows && npm run clean && cd ../..",
+    "build:all": "npm run build-common && npm run build && npm run build:api && npm run build:dataflows",
     "build": "tsc --build tsconfig.build.json",
     "build-common": "cd ../common && npm ci && npm run build && cd ../backend",
     "build:api": "cd function-apps/api && npm run build && cd ../..",
-    "build:pipelines": "cd function-apps/pipelines && npm run build && cd ../..",
-    "pack:all": "npm run pack:api && npm run pack:pipelines",
+    "build:dataflows": "cd function-apps/dataflows && npm run build && cd ../..",
+    "pack:all": "npm run pack:api && npm run pack:dataflows",
     "pack": "npm run pack:api",
     "pack:api": "cd function-apps/api && npm run pack && cd ../..",
-    "pack:pipelines": "cd function-apps/pipelines && npm run pack && cd ../..",
+    "pack:dataflows": "cd function-apps/dataflows && npm run pack && cd ../..",
     "start": "npm run start:api",
     "start:api": "cd function-apps/api && npm run start && cd ../..",
-    "start:pipelines": "cd function-apps/pipelines && npm run start && cd ../..",
+    "start:dataflows": "cd function-apps/dataflows && npm run start && cd ../..",
     "test": "CAMS_LOGIN_PROVIDER='mock' DATABASE_MOCK='true' MONGO_CONNECTION_STRING='mongodb://test-string' jest -c jest.config.js",
     "test:open-handles": "CAMS_LOGIN_PROVIDER='mock' DATABASE_MOCK='true' MONGO_CONNECTION_STRING='mongodb://test-string' jest -c jest.config.js --detectOpenHandles",
     "lint": "eslint .",
